[#1089] Paraview Nexus widget
git-svn-id: https://ccpforge.cse.rl.ac.uk/svn/iqa/trunk@137 1f895c63-7dff-42ab-8404-156523f462c5

+++ Core/NexusWidget/CCPiNexusWidgetDialog.h
@@ -18,7 +18,7 @@ public:
 	~CCPiNexusWidgetDialog();
 	void SetModel(QAbstractItemModel *);
 	std::vector<std::string> GetSelectedDataSetList(){return SelectedDatasetList;}
-	int  GetSelectedDataSetCount(){return SelectedDatasetList.size();}
+	int  GetSelectedDataSetCount(){return (int)SelectedDatasetList.size();}
 	std::string GetSelectedDataSet(int index){return SelectedDatasetList.at(index);}
 private:
 	void accept();
